Refactor BookmarkAnimeCommand into BookmarkParser for preprocessing
