+++ src/main/java/seedu/duke/parser/Parser.java
@@ -45,7 +45,7 @@ public class Parser {
             return new AddToWatchlistCommand(description);
 
         case "bookmark":
-            return new BookmarkAnimeCommand(description);
+            return new BookmarkParser().parse(description);
 
         case "help":
             return new HelpCommand();
